Fixed tests, implemented TranslatorsController tests

+++ Zotero/Controllers/API/Requests/TranslatorsRequest.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 struct TranslatorsRequest: ApiRequest {
-    let timestamp: Int
+    let timestamp: Double
     let version: String
     let type: Int
